moved list spacing decorator to common module
